requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
SpeechRecognition>=3.8.1
pydub>=0.25.1
openai>=1.0.0
python-dotenv>=0.19.0
